refactor: separate drawRaw
